# PERF: RangeIndex.format performance