--- a/doc/source/whatsnew/v1.1.2.rst
+++ b/doc/source/whatsnew/v1.1.2.rst
@@ -15,7 +15,7 @@ including other versions of pandas.
 Fixed regressions
 ~~~~~~~~~~~~~~~~~
 
--
+- Performance regression for :meth:`RangeIndex.format` (:issue:`35712`)
 -
 
 .. ---------------------------------------------------------------------------