--- a/doc/source/whatsnew/v1.1.2.rst
+++ b/doc/source/whatsnew/v1.1.2.rst
@@ -14,10 +14,10 @@ including other versions of pandas.
 
 Fixed regressions
 ~~~~~~~~~~~~~~~~~
-- Performance regression for :meth:`RangeIndex.format` (:issue:`35712`)
 - Regression in :meth:`DatetimeIndex.intersection` incorrectly raising ``AssertionError`` when intersecting against a list (:issue:`35876`)
 - Performance regression for :meth:`RangeIndex.format` (:issue:`35712`)
 -
+
 
 .. ---------------------------------------------------------------------------