--- a/pandas/core/indexes/range.py
+++ b/pandas/core/indexes/range.py
@@ -187,7 +187,7 @@ class RangeIndex(Int64Index):
         # we are formatting thru the attributes
         return None
 
-    def _format_with_header(self, header, na_rep="NaN") -> List[str]:
+    def _format_with_header(self, header: List[str], na_rep: str = "NaN") -> List[str]:
         first_val_str = str(self._range[0])
         last_val_str = str(self._range[-1])
         max_length = max(len(first_val_str), len(last_val_str))