--- a/pandas/core/indexes/range.py
+++ b/pandas/core/indexes/range.py
@@ -188,7 +188,7 @@ class RangeIndex(Int64Index):
         return None
 
     def _format_with_header(self, header: List[str], na_rep: str = "NaN") -> List[str]:
-        if len(self._range) == 0:
+        if not len(self._range):
             return header
         first_val_str = str(self._range[0])
         last_val_str = str(self._range[-1])