@jreback 2020-08-19
looks good. this is worth a perf mention in 1.2 whatsnew. ping on green.