@topper-123 2020-08-20
The tracisCI issue is just a timeout issue, but In `CI / Web and docs`, I've got a problem under  `Check ipython directive errors` that I'm not able to fix:

```
Run ! grep -B1 "^<<<-------------------------------------------------------------------------$" sphinx.log
[0 rows x 1 columns]
<<<-------------------------------------------------------------------------
##[error]Process completed with exit code 1.
```

The source code for this in `.github/workflows/ci.yml` says:

```yml
    # This can be removed when the ipython directive fails when there are errors,
    # including the `tee sphinx.log` in te previous step (https://github.com/ipython/ipython/issues/11547)
    - name: Check ipython directive errors
      run: "! grep -B1 \"^<<<-------------------------------------------------------------------------$\" sphinx.log"
```

I don't understand what this check is for. @TomAugspurger , you posted that mentioned ipython issue, do you mind looking at why my PR fails on this?