@simonjayhawkins 2020-08-26
> minor comment, if you can address and merge on green.

comment addressed