@simonjayhawkins 2020-08-26
@meeseeksdev backport 1.1.x